Rename prompt_utils.py to llm_query_processing.py
